+++ src/main/java/com/maxmind/geoip2/model/AsnResponse.java
@@ -37,9 +37,9 @@ public class AsnResponse extends AbstractResponse {
         @JsonProperty("autonomous_system_organization")
         @MaxMindDbParameter(name = "autonomous_system_organization")
         String autonomousSystemOrganization,
-        @JacksonInject("ip_address") @JsonProperty("ip_address")
+        @JsonProperty("ip_address")
         @MaxMindDbParameter(name = "ip_address") String ipAddress,
-        @JacksonInject("network") @JsonProperty("network")
+        @JsonProperty("network")
         @JsonDeserialize(using = NetworkDeserializer.class) @MaxMindDbParameter(name = "network")
         Network network
     ) {